Modified object dragging feature
Allows object to go back to original position and rotation once mouse is lifted.
