--- Assets/Scripts/Sohum/pourFluid.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d98fd8a65f0f34164afeb346cb900053
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
